# docs/api_document/booking/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CAB System - booking API",
    "version": "2.1.0",
    "description": "Hợp đồng thiết kế theo SRS X và kiến trúc 7 service. Mọi request từ client đi qua Gateway; dùng prefix /api/v1. Chưa xác nhận backend đã triển khai. Operation x-contract=extension là đề xuất bổ sung, không phải API nghiệm thu đã chốt."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Gateway demo; đổi sang địa chỉ Gateway thực tế khi triển khai"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v1/bookings": {
      "post": {
        "operationId": "booking_post__booking",
        "tags": [
          "booking"
        ],
        "summary": "Tạo yêu cầu đặt xe",
        "description": "Xác minh Customer/userId phía server. Ghi Booking/outbox rồi matching bất đồng bộ. Giới hạn 30 request/phút/user và 120/phút/IP; vượt ngưỡng trả 429 trước tạo Booking.",
        "x-service-owner": "booking",
        "x-contract": "srs",
        "x-fr": [
          "FR08"
        ],
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "bookingId",
                    "status"
                  ],
                  "properties": {
                    "bookingId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "minLength": 1,
                      "enum": [
                        "SEARCHING_DRIVER"
                      ]
                    }
                  }
                },
                "example": {
                  "bookingId": "BKG-001",
                  "status": "SEARCHING_DRIVER"
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Xác thực/chữ ký thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sai role, caller hoặc ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Vượt giới hạn gửi/request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Số giây chờ trước lần thử kế tiếp",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency chưa sẵn sàng; không suy diễn mutation chưa thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "CUSTOMER"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pickup",
                  "destination",
                  "vehicleTypeId"
                ],
                "properties": {
                  "pickup": {
                    "$ref": "#/components/schemas/Point"
                  },
                  "destination": {
                    "$ref": "#/components/schemas/Point"
                  },
                  "vehicleTypeId": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              },
              "example": {
                "pickup": {
                  "lat": 10.7769,
                  "lng": 106.7009,
                  "address": "Điểm đón demo"
                },
                "destination": {
                  "lat": 10.782,
                  "lng": 106.695,
                  "address": "Điểm đến demo"
                },
                "vehicleTypeId": "VT-001"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "booking_get__customers_id_bookings",
        "tags": [
          "booking"
        ],
        "summary": "Danh sách Booking của Customer",
        "description": "Customer chính chủ hoặc Admin/Operator; mọi trạng thái Booking. Sort createdAt DESC rồi bookingId, total sau lọc; không thay bằng lịch sử Trip.",
        "x-service-owner": "booking",
        "x-contract": "srs",
        "x-fr": [
          "FR80"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BookingStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "page",
                    "limit",
                    "total"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Booking"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 100
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Xác thực/chữ ký thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sai role, caller hoặc ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy tài nguyên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency chưa sẵn sàng; không suy diễn mutation chưa thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "CUSTOMER",
          "ADMIN",
          "OPERATOR"
        ]
      }
    },
    "/api/v1/bookings/{id}": {
      "patch": {
        "operationId": "booking_post__bookings_id_cancel",
        "tags": [
          "booking"
        ],
        "summary": "Hủy Booking/chuyến trước khi đón",
        "description": "Chính chủ, reason không rỗng. Nếu có Trip, Trip phân xử cancel/PICKED_UP bằng conditional update. Chỉ trả 200 khi Booking và Trip (nếu có) xác nhận CANCELED; Driver release và notification qua event. Hủy muộn trả 409.",
        "x-service-owner": "booking",
        "x-contract": "srs",
        "x-fr": [
          "FR10",
          "FR66"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "bookingId",
                    "status"
                  ],
                  "properties": {
                    "bookingId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "minLength": 1,
                      "enum": [
                        "CANCELED"
                      ]
                    },
                    "tripId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "tripStatus": {
                      "type": "string",
                      "minLength": 1,
                      "enum": [
                        "CANCELED"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Xác thực/chữ ký thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sai role, caller hoặc ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy tài nguyên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, dữ liệu trùng hoặc idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency chưa sẵn sàng; không suy diễn mutation chưa thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Timeout upstream; tra cứu/retry bằng cùng ID trước khi bù",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "CUSTOMER"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason",
                  "status"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "\\S"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "CANCELED"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booking-offers": {
      "get": {
        "operationId": "booking_get__drivers_me_offers",
        "tags": [
          "booking"
        ],
        "summary": "Danh sách Offer của Driver",
        "description": "Chỉ Offer của Driver đang đăng nhập; status mặc định PENDING, trả expiresAt và thông tin điểm đón/đến cần thiết.",
        "x-service-owner": "booking",
        "x-contract": "srs",
        "x-fr": [
          "FR17"
        ],
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "enum": [
                "PENDING",
                "ACCEPTED",
                "REJECTED",
                "TIMEOUT",
                "CANCELED"
              ],
              "default": "PENDING"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "page",
                    "limit",
                    "total"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Offer"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "limit": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 100
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Xác thực/chữ ký thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sai role, caller hoặc ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency chưa sẵn sàng; không suy diễn mutation chưa thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "DRIVER"
        ]
      }
    },
    "/api/v1/booking-offers/{id}": {
      "patch": {
        "operationId": "booking_patch__offers_id",
        "tags": [
          "booking"
        ],
        "summary": "Cập nhật trạng thái Offer",
        "description": "Driver được mời; Offer còn hạn. Workflow reserve Driver → tạo Trip → confirm BUSY → activate Trip. Một assignment thắng/Booking; timeout tra cứu cùng ID. Retry sau khi đã hoàn tất trả 409 theo SRS.",
        "x-service-owner": "booking",
        "x-contract": "srs",
        "x-fr": [
          "FR21"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "tripId",
                    "bookingId",
                    "driverId",
                    "status"
                  ],
                  "properties": {
                    "tripId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "bookingId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "driverId": {
                      "type": "string",
                      "minLength": 1
                    },
                    "status": {
                      "type": "string",
                      "minLength": 1,
                      "enum": [
                        "ASSIGNED"
                      ]
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "Thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Xác thực/chữ ký thất bại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Sai role, caller hoặc ownership",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy tài nguyên",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Xung đột trạng thái, dữ liệu trùng hoặc idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency chưa sẵn sàng; không suy diễn mutation chưa thực hiện",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Timeout upstream; tra cứu/retry bằng cùng ID trước khi bù",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-roles": [
          "DRIVER"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACCEPTED",
                      "REJECTED"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT người dùng; verify signature/issuer/audience/exp và ownership ở service."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "correlationId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "correlationId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Point": {
        "type": "object",
        "required": [
          "lat",
          "lng",
          "address"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "lng": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "address": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "BookingStatus": {
        "type": "string",
        "minLength": 1,
        "enum": [
          "CREATED",
          "SEARCHING_DRIVER",
          "DRIVER_ASSIGNED",
          "CONFIRMED",
          "COMPLETED",
          "CANCELED",
          "NO_DRIVER_FOUND"
        ]
      },
      "Booking": {
        "type": "object",
        "required": [
          "bookingId",
          "customerId",
          "pickup",
          "destination",
          "vehicleTypeId",
          "status",
          "createdAt"
        ],
        "properties": {
          "bookingId": {
            "type": "string",
            "minLength": 1
          },
          "customerId": {
            "type": "string",
            "minLength": 1
          },
          "pickup": {
            "$ref": "#/components/schemas/Point"
          },
          "destination": {
            "$ref": "#/components/schemas/Point"
          },
          "vehicleTypeId": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus"
          },
          "createdAt": {
            "type": "string",
            "minLength": 1,
            "format": "date-time"
          },
          "tripId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Offer": {
        "type": "object",
        "required": [
          "offerId",
          "bookingId",
          "driverId",
          "status",
          "expiresAt",
          "pickup",
          "destination"
        ],
        "properties": {
          "offerId": {
            "type": "string",
            "minLength": 1
          },
          "bookingId": {
            "type": "string",
            "minLength": 1
          },
          "driverId": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "TIMEOUT",
              "CANCELED"
            ]
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "format": "date-time"
          },
          "pickup": {
            "$ref": "#/components/schemas/Point"
          },
          "destination": {
            "$ref": "#/components/schemas/Point"
          }
        }
      }
    }
  },
  "x-service-owner": "booking",
  "x-persistence": {
    "engine": "PostgreSQL",
    "database": "booking_db",
    "description": "Metadata thiết kế; contract HTTP không phải schema vật lý. Mỗi service chỉ truy cập database của mình."
  }
}